Add Year 1 Web Development Fundamentals - Personal Portfolio project
This is my Personal Portfolio project created for the **Web Development Fundamentals** module.
The goal of this assessment was to build a fully functional, responsive, multi-page website that demonstrates the core HTML, CSS, and JavaScript skills developed throughout the semester.

The project was created following the official assignment brief, which required multiple pages, proper navigation, consistent styling, and adherence to web development best practices.

Do not at the commit date, the deadline was 10th of December 2023, so I complete this project at the time. I am just reorganizing the way this repo looks and i decided to put this project on a new branch