Add Year 2 GUI Programming - Final Assessment (version 8)
Introduces two new Java Swing applications: HouseholdCooker, a UI for controlling cooking zones with adjustable intensity and font size options, and MouseEvents, which displays an image on mouse hover. Includes supporting images for both applications.

This is a 1 day assignment, deadline:
23:59pm on December 5th 2024.
That what I has done, but uploading to this branch late, and I don't want to change the date on the commit...