Add Year 2 Advanced Programming - Final Assessment (version 5)
Added log detail files, instrument images, sound files, internationalization resources, class diagram sketch, and source code for exceptions, GUI, instruments, internationalization, and logging. This commit introduces the initial assets and implementation for the musical instruments project.

This is a 1 day assignment (3rd April 2025), deadline:
23:59pm on April 3rd 2025.
That what I has done, but uploading to this branch late, and I don't want to change the date on the commit...