Add Year 1 Algorithmic Problem Solving - In-lab practical test
Added Java code and pseudocode for question 1 in the submission folder. Also included a PDF solution for question 2.

CA1 was just one question on which  we should been write pseoudocode & code.
CA2 was to answer on algorithmic questions.

This was a long time ago on a lab session (2023), but i am uploading to here now.